Refactor models & tests
