+++ test/models/launch_test.dart
@@ -624,296 +624,6 @@ void main() {
     });
   });
 
-  group('Core', () {
-    test('is correctly generated from a JSON', () {
-      expect(
-        Core.fromJson(const {
-          "core": {
-            "block": 5,
-            "reuse_count": 0,
-            "rtls_attempts": 0,
-            "rtls_landings": 0,
-            "asds_attempts": 0,
-            "asds_landings": 0,
-            "last_update":
-                "Engine failure at T+33 seconds resulted in loss of vehicle",
-            "launches": [
-              {
-                "flight_number": 1,
-                "name": "FalconSat",
-                "date_utc": "2015-06-28T14:21:00.000Z",
-                "id": "5eb87cd9ffd86e000604b32a"
-              }
-            ],
-            "serial": "Merlin1A",
-            "status": "lost",
-            "id": "5e9e289df35918033d3b2623"
-          },
-          "gridfins": false,
-          "legs": false,
-          "reused": false,
-          "landing_attempt": false,
-          "landing_success": false,
-          "landpad": {
-            "name": "OCISLY",
-            "full_name": "Of Course I Still Love You",
-            "type": "ASDS",
-            "locality": "Port Canaveral",
-            "region": "Florida",
-            "latitude": 28.4104,
-            "longitude": -80.6188,
-            "landing_attempts": 36,
-            "landing_successes": 30,
-            "wikipedia":
-                "https://en.wikipedia.org/wiki/Autonomous_spaceport_drone_ship",
-            "details":
-                "The second ASDS barge, Of Course I Still Love You (OCISLY), had been under construction in a Louisiana shipyard since early 2015 using a different hull—Marmac 304—in order to service launches on the east coast. It was built as a replacement for the first Just Read the Instructions and entered operational service for Falcon 9 Flight 19 in late June 2015. As of June 2015, its home port was Jacksonville, Florida, but after December 2015, it was transferred 160 miles (260 km) further south, at Port Canaveral. While the dimensions of the ship are nearly identical to the first ASDS, several enhancements were made including a steel blast wall erected between the aft containers and the landing deck. The ship was in place for a first-stage landing test on the CRS-7 mission, which failed on launch on 28 June 2015. On 8 April 2016 the first stage, which launched the Dragon CRS-8 spacecraft, successfully landed for the first time ever on OCISLY, which is also the first ever drone ship landing. In February 2018, the Falcon Heavy Test Flight's central core exploded upon impact next to OCISLY that damaged two of the four thrusters on the drone ship. Two thrusters were removed from the Marmac 303 barge in order to repair OCISLY.",
-            "status": "active",
-            "id": "5e9e3032383ecb6bb234e7ca"
-          }
-        }),
-        Core(
-          block: 5,
-          reuseCount: 0,
-          rtlsAttempts: 0,
-          rtlsLandings: 0,
-          asdsAttempts: 0,
-          asdsLandings: 0,
-          lastUpdate:
-              'Engine failure at T+33 seconds resulted in loss of vehicle',
-          launches: [
-            LaunchDetails(
-              flightNumber: 1,
-              name: 'FalconSat',
-              date: DateTime.parse('2015-06-28T14:21:00.000Z'),
-              id: '5eb87cd9ffd86e000604b32a',
-            ),
-          ],
-          serial: 'Merlin1A',
-          status: 'lost',
-          id: '5e9e289df35918033d3b2623',
-          hasGridfins: false,
-          hasLegs: false,
-          reused: false,
-          landingAttempt: false,
-          landingSuccess: false,
-          landpad: LandpadDetails(
-            name: 'OCISLY',
-            fullName: 'Of Course I Still Love You',
-            type: 'ASDS',
-            locality: 'Port Canaveral',
-            region: 'Florida',
-            latitude: 28.4104,
-            longitude: -80.6188,
-            landingAttempts: 36,
-            landingSuccesses: 30,
-            wikipediaUrl:
-                'https://en.wikipedia.org/wiki/Autonomous_spaceport_drone_ship',
-            details:
-                "The second ASDS barge, Of Course I Still Love You (OCISLY), had been under construction in a Louisiana shipyard since early 2015 using a different hull—Marmac 304—in order to service launches on the east coast. It was built as a replacement for the first Just Read the Instructions and entered operational service for Falcon 9 Flight 19 in late June 2015. As of June 2015, its home port was Jacksonville, Florida, but after December 2015, it was transferred 160 miles (260 km) further south, at Port Canaveral. While the dimensions of the ship are nearly identical to the first ASDS, several enhancements were made including a steel blast wall erected between the aft containers and the landing deck. The ship was in place for a first-stage landing test on the CRS-7 mission, which failed on launch on 28 June 2015. On 8 April 2016 the first stage, which launched the Dragon CRS-8 spacecraft, successfully landed for the first time ever on OCISLY, which is also the first ever drone ship landing. In February 2018, the Falcon Heavy Test Flight's central core exploded upon impact next to OCISLY that damaged two of the four thrusters on the drone ship. Two thrusters were removed from the Marmac 303 barge in order to repair OCISLY.",
-            status: 'active',
-            id: '5e9e3032383ecb6bb234e7ca',
-          ),
-        ),
-      );
-    });
-  });
-
-  group('Crew', () {
-    test('is correctly generated from a JSON', () {
-      expect(
-        Crew.fromJson(const {
-          "name": "Douglas Hurley",
-          "agency": "NASA",
-          "image": "https://i.imgur.com/ooaayWf.png",
-          "wikipedia": "https://en.wikipedia.org/wiki/Douglas_G._Hurley",
-          "launches": [
-            {
-              "flight_number": 94,
-              "name": "CCtCap Demo Mission 2",
-              "date_utc": "2015-06-28T14:21:00.000Z",
-              "id": "5eb87d46ffd86e000604b388"
-            }
-          ],
-          "status": "active",
-          "id": "5ebf1b7323a9a60006e03a7b"
-        }),
-        Crew(
-          name: 'Douglas Hurley',
-          agency: 'NASA',
-          imageUrl: 'https://i.imgur.com/ooaayWf.png',
-          wikipediaUrl: 'https://en.wikipedia.org/wiki/Douglas_G._Hurley',
-          launches: [
-            LaunchDetails(
-              flightNumber: 94,
-              name: 'CCtCap Demo Mission 2',
-              date: DateTime.parse('2015-06-28T14:21:00.000Z'),
-              id: '5eb87d46ffd86e000604b388',
-            ),
-          ],
-          status: 'active',
-          id: '5ebf1b7323a9a60006e03a7b',
-        ),
-      );
-    });
-  });
-
-  group('Payload', () {
-    test('is correctly generated from a JSON', () {
-      expect(
-        Payload.fromJson(const {
-          "dragon": {
-            "capsule": {
-              "reuse_count": 1,
-              "water_landings": 1,
-              "last_update": "descripction",
-              "launches": [
-                {
-                  "flight_number": 10,
-                  "name": "CRS-2",
-                  "date_utc": "2015-06-28T14:21:00.000Z",
-                  "id": "5eb87ce1ffd86e000604b333"
-                }
-              ],
-              "serial": "C104",
-              "status": "unknown",
-              "id": "5e9e2c5bf359189ef23b2667"
-            }
-          },
-          "name": "FalconSAT-2",
-          "reused": true,
-          "customers": ["DARPA"],
-          "nationalities": ["United States"],
-          "manufacturers": ["SSTL"],
-          "mass_kg": 20,
-          "orbit": "LEO",
-          "periapsis_km": 400,
-          "apoapsis_km": 500,
-          "inclination_deg": 39,
-          "period_min": 90.0,
-          "id": "5eb0e4b5b6c3bb0006eeb1e1"
-        }),
-        Payload(
-          capsule: CapsuleDetails(
-            reuseCount: 1,
-            splashings: 1,
-            lastUpdate: 'descripction',
-            launches: [
-              LaunchDetails(
-                flightNumber: 10,
-                name: 'CRS-2',
-                date: DateTime.parse('2015-06-28T14:21:00.000Z'),
-                id: '5eb87ce1ffd86e000604b333',
-              )
-            ],
-            serial: 'C104',
-            status: 'unknown',
-            id: '5e9e2c5bf359189ef23b2667',
-          ),
-          name: 'FalconSAT-2',
-          reused: true,
-          customer: 'DARPA',
-          nationality: 'United States',
-          manufacturer: 'SSTL',
-          mass: 20,
-          orbit: 'LEO',
-          periapsis: 400,
-          apoapsis: 500,
-          inclination: 39,
-          period: 90.0,
-          id: '5eb0e4b5b6c3bb0006eeb1e1',
-        ),
-      );
-    });
-  });
-
-  group('LaunchpadDetails', () {
-    test('is correctly generated from a JSON', () {
-      expect(
-        LaunchpadDetails.fromJson(const {
-          "name": "Kwajalein Atoll",
-          "full_name": "Kwajalein Atoll Omelek Island",
-          "locality": "Omelek Island",
-          "region": "Marshall Islands",
-          "latitude": 9.0477206,
-          "longitude": 167.7431292,
-          "launch_attempts": 5,
-          "launch_successes": 2,
-          "details":
-              "SpaceX's primary Falcon 9 pad, where all east coast Falcon 9s launched prior to the AMOS-6 anomaly. Previously used alongside SLC-41 to launch Titan rockets for the US Air Force, the pad was heavily damaged by the AMOS-6 anomaly in September 2016. It returned to flight with CRS-13 on December 15, 2017, boasting an upgraded throwback-style Transporter-Erector modeled after that at LC-39A.",
-          "id": "5e9e4502f5090995de566f86"
-        }),
-        LaunchpadDetails(
-          name: 'Kwajalein Atoll',
-          fullName: 'Kwajalein Atoll Omelek Island',
-          locality: 'Omelek Island',
-          region: 'Marshall Islands',
-          latitude: 9.0477206,
-          longitude: 167.7431292,
-          launchAttempts: 5,
-          launchSuccesses: 2,
-          details:
-              "SpaceX's primary Falcon 9 pad, where all east coast Falcon 9s launched prior to the AMOS-6 anomaly. Previously used alongside SLC-41 to launch Titan rockets for the US Air Force, the pad was heavily damaged by the AMOS-6 anomaly in September 2016. It returned to flight with CRS-13 on December 15, 2017, boasting an upgraded throwback-style Transporter-Erector modeled after that at LC-39A.",
-          id: '5e9e4502f5090995de566f86',
-        ),
-      );
-    });
-  });
-
-  group('CapsuelDetails', () {
-    test('is correctly generated from a JSON', () {
-      expect(
-        CapsuleDetails.fromJson(const {
-          'reuse_count': 1,
-          'water_landings': 1,
-          'last_update': 'descripction',
-          'launches': [
-            {
-              'flight_number': 10,
-              'name': 'CRS-2',
-              "date_utc": "2015-06-28T14:21:00.000Z",
-              'id': '5eb87ce1ffd86e000604b333'
-            }
-          ],
-          'serial': 'C104',
-          'status': 'unknown',
-          'id': '5e9e2c5bf359189ef23b2667'
-        }),
-        CapsuleDetails(
-          reuseCount: 1,
-          splashings: 1,
-          lastUpdate: 'descripction',
-          launches: [
-            LaunchDetails(
-              flightNumber: 10,
-              name: 'CRS-2',
-              date: DateTime.parse('2015-06-28T14:21:00.000Z'),
-              id: '5eb87ce1ffd86e000604b333',
-            )
-          ],
-          serial: 'C104',
-          status: 'unknown',
-          id: '5e9e2c5bf359189ef23b2667',
-        ),
-      );
-    });
-  });
-
-  group('ShipDetails', () {
-    test('is correctly generated from a JSON', () {
-      expect(
-        ShipDetails.fromJson(const {
-          'name': 'OCISLY',
-          'id': '5e9e3032383ecb6bb234e7ca',
-        }),
-        ShipDetails(
-          name: 'OCISLY',
-          id: '5e9e3032383ecb6bb234e7ca',
-        ),
-      );
-    });
-  });
-
   group('FailureDetails', () {
     test('is correctly generated from a JSON', () {
       expect(
@@ -928,47 +638,6 @@ void main() {
           altitude: 40,
           reason:
               "helium tank overpressure lead to the second stage LOX tank explosion",
-        ),
-      );
-    });
-  });
-
-  group('LandpadDetails', () {
-    test('is correctly generated from a JSON', () {
-      expect(
-        LandpadDetails.fromJson(const {
-          'name': 'OCISLY',
-          'full_name': 'Of Course I Still Love You',
-          'type': 'ASDS',
-          'locality': 'Port Canaveral',
-          'region': 'Florida',
-          'latitude': 28.4104,
-          'longitude': -80.6188,
-          'landing_attempts': 36,
-          'landing_successes': 30,
-          'wikipedia':
-              'https://en.wikipedia.org/wiki/Autonomous_spaceport_drone_ship',
-          'details':
-              "The second ASDS barge, Of Course I Still Love You (OCISLY), had been under construction in a Louisiana shipyard since early 2015 using a different hull—Marmac 304—in order to service launches on the east coast. It was built as a replacement for the first Just Read the Instructions and entered operational service for Falcon 9 Flight 19 in late June 2015. As of June 2015, its home port was Jacksonville, Florida, but after December 2015, it was transferred 160 miles (260 km) further south, at Port Canaveral. While the dimensions of the ship are nearly identical to the first ASDS, several enhancements were made including a steel blast wall erected between the aft containers and the landing deck. The ship was in place for a first-stage landing test on the CRS-7 mission, which failed on launch on 28 June 2015. On 8 April 2016 the first stage, which launched the Dragon CRS-8 spacecraft, successfully landed for the first time ever on OCISLY, which is also the first ever drone ship landing. In February 2018, the Falcon Heavy Test Flight's central core exploded upon impact next to OCISLY that damaged two of the four thrusters on the drone ship. Two thrusters were removed from the Marmac 303 barge in order to repair OCISLY.",
-          'status': 'active',
-          'id': '5e9e3032383ecb6bb234e7ca'
-        }),
-        LandpadDetails(
-          name: 'OCISLY',
-          fullName: 'Of Course I Still Love You',
-          type: 'ASDS',
-          locality: 'Port Canaveral',
-          region: 'Florida',
-          latitude: 28.4104,
-          longitude: -80.6188,
-          landingAttempts: 36,
-          landingSuccesses: 30,
-          wikipediaUrl:
-              'https://en.wikipedia.org/wiki/Autonomous_spaceport_drone_ship',
-          details:
-              "The second ASDS barge, Of Course I Still Love You (OCISLY), had been under construction in a Louisiana shipyard since early 2015 using a different hull—Marmac 304—in order to service launches on the east coast. It was built as a replacement for the first Just Read the Instructions and entered operational service for Falcon 9 Flight 19 in late June 2015. As of June 2015, its home port was Jacksonville, Florida, but after December 2015, it was transferred 160 miles (260 km) further south, at Port Canaveral. While the dimensions of the ship are nearly identical to the first ASDS, several enhancements were made including a steel blast wall erected between the aft containers and the landing deck. The ship was in place for a first-stage landing test on the CRS-7 mission, which failed on launch on 28 June 2015. On 8 April 2016 the first stage, which launched the Dragon CRS-8 spacecraft, successfully landed for the first time ever on OCISLY, which is also the first ever drone ship landing. In February 2018, the Falcon Heavy Test Flight's central core exploded upon impact next to OCISLY that damaged two of the four thrusters on the drone ship. Two thrusters were removed from the Marmac 303 barge in order to repair OCISLY.",
-          status: 'active',
-          id: '5e9e3032383ecb6bb234e7ca',
         ),
       );
     });
